% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{basic.lmm.analysis}
\alias{basic.lmm.analysis}
\title{Performs a simple lmm analysis to illustrate how to compute
confidence intervals of means.
Need to switch over to emmeans.
WARNING - this function invokes bayesian analysis with MCMC}
\usage{
basic.lmm.analysis()
}
\description{
Performs a simple lmm analysis to illustrate how to compute
confidence intervals of means.
Need to switch over to emmeans.
WARNING - this function invokes bayesian analysis with MCMC
}
\seealso{
\url{https://cran.r-project.org/web/packages/emmeans/index.html}
basic.lmm.analysis()
}
