% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{basic.lmm.dat}
\alias{basic.lmm.dat}
\title{Creates three period temporal data
with two arms.
Difference at baseline is 3 units
Difference at time 1 is 5 units
Difference at time 2 is 15 units}
\usage{
basic.lmm.dat(n.per.timept = 100, plot = T)
}
\arguments{
\item{n.per.timept}{number of points per timept (even number)}

\item{plot}{plot data if T}
}
\description{
Creates three period temporal data
with two arms.
Difference at baseline is 3 units
Difference at time 1 is 5 units
Difference at time 2 is 15 units
}
\examples{
basic.lmm.dat()
}
