% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{Cbind}
\alias{Cbind}
\title{Version of cbind to use when you have different length data.frames}
\usage{
Cbind(..., first = TRUE)
}
\description{
Version of cbind to use when you have different length data.frames
}
\examples{
Cbind()
}
\keyword{cats}
