% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{ci.logistic}
\alias{ci.logistic}
\title{CI for logistic regression model (for odds ratios)}
\usage{
ci.logistic(lm, dp = 3)
}
\description{
CI for logistic regression model (for odds ratios)
}
\examples{
ci.logistic()
}
\keyword{cats}
