% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{col.names}
\alias{col.names}
\title{Length of unique values}
\usage{
col.names(df)
}
\arguments{
\item{df}{Vector of values.
col.names()}
}
\description{
This function gives you the number of unique values in a vector.
}
