% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{corstarsl}
\alias{corstarsl}
\title{Codes p-values into significance stars}
\usage{
corstarsl(x)
}
\arguments{
\item{x}{Vector of p-values}
}
\description{
Translates p < 0.05 to * etc.
}
\examples{
corstarsl(x)
}
