% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{desc.stat.str}
\alias{desc.stat.str}
\title{Descriptive stats as a latex formatted string}
\usage{
desc.stat.str(x)
}
\arguments{
\item{x}{Vector of values.}
}
\description{
This function gives you the number of unique values in a vector.
}
\examples{
desc.stat.str()
}
\keyword{cats}
