% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{diff_means}
\alias{diff_means}
\title{Same as the ttesti in stata (for unequal var and with welch approx)}
\usage{
diff_means(x1, x2, s1, s2, n1, n2)
}
\arguments{
\item{x1}{Vector of 1st sample means}

\item{x2}{Vector of 2nd sample means}

\item{s1}{Vector of 1st sample sd}

\item{s2}{Vector of 2nd sample sd}

\item{n1}{Vector of 1st sample size}

\item{n2}{Vector of 2nd sample size}
}
\description{
looks at x2 - x1
}
\examples{
diff_means()
}
\keyword{cats}
