% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{dof_approx}
\alias{dof_approx}
\title{Welch satterthwaite - degrees of freedom approximation}
\usage{
dof_approx(s1, s2, n1, n2)
}
\arguments{
\item{s1}{Vector of 1st sample sd}

\item{s2}{Vector of 2nd sample sd}

\item{n1}{Vector of 1st sample size}

\item{n2}{Vector of 2nd sample size
dof_approx()}
}
\description{
Welch satterthwaite - degrees of freedom approximation
}
