% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{duplicate_df}
\alias{duplicate_df}
\title{Duplicates a dataframe n times and appends the duplicates
to the original one.}
\usage{
duplicate_df(df, times = 2)
}
\arguments{
\item{df}{data frame to duplicate}

\item{times}{number of copies required}
}
\description{
Duplicates a dataframe n times and appends the duplicates
to the original one.
}
\examples{
duplicate_df()
}
\keyword{emma}
