% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{expss.tab.example}
\alias{expss.tab.example}
\title{Various expss examples.}
\usage{
expss.tab.example()
}
\description{
Various expss examples.
}
\seealso{
\url{https://cran.r-project.org/web/packages/expss/vignettes/tables-with-labels.html}
expss.tab.example()
}
