% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{freq_prop}
\alias{freq_prop}
\title{Provides frequency and proportion of a given value in a vector of values}
\usage{
freq_prop(x, level, dp = 1, percent = T)
}
\arguments{
\item{x}{vector of values}

\item{level}{entry of interest (can be NA)}

\item{dp}{decimal places}

\item{percent}{provide in percentage terms}
}
\description{
Provides frequency and proportion of a given value in a vector of values
}
\examples{
freq_prop()
}
\keyword{fred}
