% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{gg_testplot01}
\alias{gg_testplot01}
\title{Just reference code.
Produces a list of ggplots with different font styles.}
\usage{
gg_testplot01()
}
\description{
Just reference code.
Produces a list of ggplots with different font styles.
}
\examples{
gg_testplot01()
}
\keyword{emma}
