% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{glm.check.res.dev}
\alias{glm.check.res.dev}
\title{Residual deviance check for GLM}
\usage{
glm.check.res.dev(myglm)
}
\arguments{
\item{myglm}{A fitted glm model}
}
\description{
Compares resid dev with chisq having resid.df degrees of freedom.
}
\examples{
glm.check.res.dev()
}
