% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{glm_table}
\alias{glm_table}
\title{Produces a basic latex table for glm coefs.}
\usage{
glm_table(lm1, dp = 2, inverselink = exp, outname = "tmp.tex")
}
\arguments{
\item{lm1}{glm}

\item{dp}{dec places to use}

\item{thefilename}{output filename}
}
\description{
Produces a basic latex table for glm coefs.
}
\examples{
glm_table()
}
\keyword{emma}
