% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{how_many_levels}
\alias{how_many_levels}
\title{Number of levels in a factor vector}
\usage{
how_many_levels(x)
}
\arguments{
\item{x}{Vector of p-values
how_many_levels(x)}
}
\description{
Number of levels in a factor vector
}
