% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{init}
\alias{init}
\title{Sets up ggplot themes and some other constructor like stuff.}
\usage{
init()
}
\description{
Sets up ggplot themes and some other constructor like stuff.
}
\examples{
init()
}
