% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{iqr_to_sd}
\alias{iqr_to_sd}
\title{Converts iqr to sd using a simple
normal approximation rule.}
\usage{
iqr_to_sd(q1, q3)
}
\arguments{
\item{q1}{first quartile}

\item{q3}{third quartile}
}
\description{
Converts iqr to sd using a simple
normal approximation rule.
}
\examples{
iqr_to_sd()
}
