% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{list_to_df}
\alias{list_to_df}
\title{List to data frame (actually tibble). Uses length of
first vector to define number of rows. Uses the names
of the vectors as the column names of the data.frame.}
\usage{
list_to_df(l0)
}
\arguments{
\item{l0}{list of vectors all of equal length}
}
\description{
List to data frame (actually tibble). Uses length of
first vector to define number of rows. Uses the names
of the vectors as the column names of the data.frame.
}
\examples{
list_to_df()
}
\keyword{cats}
