% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{mean.sd}
\alias{mean.sd}
\title{Means and standard deviation reported as string}
\usage{
\method{mean}{sd}(x, dp = 2)
}
\arguments{
\item{x}{Vector of p-values}
}
\description{
Means and standard deviation reported as string
}
\examples{
mean.sd()
}
