% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{mean_sd}
\alias{mean_sd}
\title{Means and standard deviation reported as string}
\usage{
mean_sd(x, dp = 2)
}
\arguments{
\item{x}{Vector of numeric values
mean_sd()}
}
\description{
Means and standard deviation reported as string
}
