% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{n.miss}
\alias{n.miss}
\title{How many are missing in my vector x}
\usage{
n.miss(x)
}
\arguments{
\item{x}{Vector of p-values
n.miss()}
}
\description{
How many are missing in my vector x
}
