% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{not_numeric}
\alias{not_numeric}
\title{Displays a sorted vector of values that were not numeric.}
\usage{
not_numeric(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Displays a sorted vector of values that were not numeric.
}
\examples{
not_numeric()
}
