% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{pad_str}
\alias{pad_str}
\title{Prepends the prefix to pad x to nchrs in length}
\usage{
pad_str(x, prefix, nchrs)
}
\arguments{
\item{x}{value to pad}

\item{prefix}{desired prefix}

\item{nchrs}{final number of chars in string}
}
\description{
Prepends the prefix to pad x to nchrs in length
}
\examples{
pad_zero()
}
