% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{padNA}
\alias{padNA}
\title{Used by Cbind}
\usage{
padNA(mydata, rowsneeded, first = TRUE)
}
\arguments{
\item{mydata}{Data}

\item{rowsneeded}{Number of padding rows}

\item{first}{Do first row}
}
\description{
Used by Cbind
}
\examples{
padNA(mydata, rowsneeded, first = TRUE)
}
