% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{plot.missing}
\alias{plot.missing}
\title{Method to call to produce ggplot of missingness 
pattern and save to file.}
\usage{
plot.missing(df, filename, showme = T)
}
\arguments{
\item{df}{Data frame}
}
\description{
Method to call to produce ggplot of missingness 
pattern and save to file.
}
\examples{
plot.missing()
}
\keyword{cats}
