% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{pooled_sd}
\alias{pooled_sd}
\title{Gets pooled standard deviation}
\usage{
pooled_sd(s1, s2, n1, n2)
}
\arguments{
\item{s1}{Vector of 1st sample sd}

\item{s2}{Vector of 2nd sample sd}

\item{n1}{Vector of 1st sample size}

\item{n2}{Vector of 2nd sample size}
}
\description{
Gets pooled standard deviation
}
\examples{
pooled_sd()
}
\keyword{cats}
