% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{predict.lmm}
\alias{predict.lmm}
\title{Bootstraps based on provided lmm and newdata data frame.
Returns the bootstrap samples enabling you to estimate
whatever you want.}
\usage{
\method{predict}{lmm}(lmm, newdata = df.new, nsim = 10, seed = 1)
}
\description{
Bootstraps based on provided lmm and newdata data frame.
Returns the bootstrap samples enabling you to estimate
whatever you want.
}
\examples{
predict.lmm()
}
