% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{prop}
\alias{prop}
\title{Proportion of entries in a vector equal to 
a specified value. Provided as a fraction or
percentage.}
\usage{
prop(x, level, dp = 1, percent = T)
}
\arguments{
\item{x}{vector of values}

\item{level}{entry of interest}

\item{dp}{decimal places}

\item{percent}{provide in percentage terms}
}
\description{
Proportion of entries in a vector equal to 
a specified value. Provided as a fraction or
percentage.
}
\examples{
list_to_df()
}
\keyword{fred}
