% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{replace.me}
\alias{replace.me}
\title{Replace a value with another value.
Useful for datasets where the null value is coded 99 etc.}
\usage{
replace.me(x, repl = 88)
}
\description{
Replace a value with another value.
Useful for datasets where the null value is coded 99 etc.
}
\examples{
replace.me()
}
\keyword{cats}
