% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{replace.p.values}
\alias{replace.p.values}
\title{Converts the p-values to the common rounded representations (< 0.05 etc).}
\usage{
replace.p.values(x, dp = 3)
}
\arguments{
\item{x}{vector of p-values}

\item{dp}{dec places to use}
}
\description{
Converts the p-values to the common rounded representations (< 0.05 etc).
}
\examples{
replace.p.values()
}
\keyword{emma}
