% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{rwedge}
\alias{rwedge}
\title{Generates a n random numbers from a wedgeshaped
distribution with lower value 0 and upper value x.}
\usage{
rwedge(n = 1, upr = 3.12)
}
\arguments{
\item{n}{number of values to produce}

\item{upr}{upper bound of wedge}
}
\description{
Generates a n random numbers from a wedgeshaped
distribution with lower value 0 and upper value x.
}
\examples{
rwedge()
}
