% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{sd.vs.se}
\alias{sd.vs.se}
\title{Standard error estimates the standard deviation of
sample means.}
\usage{
sd.vs.se()
}
\description{
Standard error estimates the standard deviation of
sample means.
}
\examples{
sd.vs.se()
}
