% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{star.wars}
\alias{star.wars}
\title{Converts the p-values to significance stars.}
\usage{
star.wars(x)
}
\arguments{
\item{x}{vector of p-values}
}
\description{
Converts the p-values to significance stars.
}
\examples{
star.wars()
}
\keyword{emma}
