% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{stone_to_kg}
\alias{stone_to_kg}
\title{Converts stone to kg.}
\usage{
stone_to_kg(x = 1)
}
\arguments{
\item{x}{value in stone to convert}
}
\description{
Converts stone to kg.
}
\examples{
stone_to_kg()
}
