% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{trim.leading}
\alias{trim.leading}
\title{returns string w/o leading whitespace}
\usage{
trim.leading(x)
}
\arguments{
\item{x}{vector of character strings}
}
\description{
returns string w/o leading whitespace
}
\examples{
trim.leading()
}
\keyword{cats}
