% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{trim.trailing}
\alias{trim.trailing}
\title{returns string w/o trailing whitespace}
\usage{
trim.trailing(x)
}
\arguments{
\item{x}{vector of character strings}
}
\description{
returns string w/o trailing whitespace
}
\examples{
trim.trailing()
}
