% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{cutpt.max.sens.spec}
\alias{cutpt.max.sens.spec}
\title{Identifies the max sensitivity and specificity cutpoints}
\usage{
cutpt.max.sens.spec(x)
}
\description{
Identifies the max sensitivity and specificity cutpoints
}
\examples{
cutpt.max.sens.spec()
}
