% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{extract.up.to.whitespace}
\alias{extract.up.to.whitespace}
\title{returns string up to the first instance of whitespace}
\usage{
extract.up.to.whitespace(x)
}
\arguments{
\item{x}{vector of character strings
extract.up.to.whitespace()}
}
\description{
returns string up to the first instance of whitespace
}
