% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{Produces ggplot of missingness pattern}
\usage{
ggplot_missing(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Produces ggplot of missingness pattern
}
\examples{
ggplot_missing()
}
