% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{len.unique}
\alias{len.unique}
\title{Length of unique values}
\usage{
len.unique(x)
}
\arguments{
\item{x}{Vector of values.}
}
\description{
This function gives you the number of unique values in a vector.
}
\examples{
len.unique()
}
\keyword{cats}
