% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{not_numeric}
\alias{not_numeric}
\title{Utility function that provides a sorted vector of the
values that were not numeric in the vector passed to
not_numeric.}
\usage{
not_numeric(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Utility function that provides a sorted vector of the
values that were not numeric in the vector passed to
not_numeric.
}
\examples{
not_numeric()
}
