% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{plot.missing}
\alias{plot.missing}
\title{Method to call to produce ggplot of missingness
pattern and save to file.}
\usage{
plot.missing(df, filename, showme = T, width = 5, height = 5)
}
\arguments{
\item{df}{Data frame}

\item{filename}{file of plot (uses Output directory as prefix)}

\item{showme}{display plot as well as save}

\item{width}{width (cm)}

\item{height}{height (cm)}
}
\description{
Method to call to produce ggplot of missingness
pattern and save to file.
}
\examples{
plot.missing()
}
\keyword{cats}
