% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{plot.missing}
\alias{plot.missing}
\title{Produces ggplot of missingness pattern and saves to file}
\usage{
plot.missing(df, filename)
}
\arguments{
\item{df}{Data frame}
}
\description{
Produces ggplot of missingness pattern and saves to file
}
\examples{
ggplot_missing()
}
