% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majutils.R
\name{trim}
\alias{trim}
\title{returns string w/o leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{vector of character strings
trim()}
}
\description{
returns string w/o leading or trailing whitespace
}
